// src/interfaces/review.interface.ts
// Generated by https://quicktype.io

export interface Review {
	id: string;
	user_id: string;
	product_id: string;
	rating: number;
	title: string;
	comment: string;
	created_at: string;
	updated_at: string;
	deleted_at: null;
	user: ReviewUser;
}

export interface ReviewUser {
	first_name: string;
	last_name: string;
	created_at: string;
}